refactored code and organized little things.
