--- resources/plugin.yml
@@ -1,4 +1,0 @@
-name: HologramDisplays
-author: Watch54
-main: fr.watch54.displays.Plugin
-version: 1.0.0
