--- src/fr/watch54/displays/Plugin.java
@@ -1,5 +1,0 @@
-package fr.watch54.displays;
-
-import org.bukkit.plugin.java.JavaPlugin;
-
-public class Plugin extends JavaPlugin {}
