--- src/fr/watch54/displays/interfaces/Text.java
@@ -1,7 +1,0 @@
-package fr.watch54.displays.interfaces;
-
-public interface Text {
-
-    String getText();
-
-}
