--- src/fr/watch54/displays/listeners/ChunkUnloadListener.java
@@ -1,15 +1,0 @@
-package fr.watch54.displays.listeners;
-
-import org.bukkit.event.EventHandler;
-import org.bukkit.event.Listener;
-import org.bukkit.event.world.ChunkUnloadEvent;
-
-public class ChunkUnloadListener implements Listener {
-
-    @EventHandler
-    public void onChunkUnload(ChunkUnloadEvent event){
-        event.setCancelled(true);
-
-    }
-
-}
