+++ src/main/java/fr/watch54/displays/interfaces/Action.java
@@ -1,0 +1,9 @@
+package fr.watch54.displays.interfaces;
+
+import org.bukkit.entity.Player;
+
+public interface Action {
+
+    void execute(Player player);
+
+}
