+++ src/main/java/fr/watch54/displays/interfaces/Text.java
@@ -1,0 +1,7 @@
+package fr.watch54.displays.interfaces;
+
+public interface Text {
+
+    String getText();
+
+}
