+++ src/main/java/resources/plugin.yml
@@ -1,0 +1,4 @@
+name: HologramDisplays
+author: Watch54
+main: fr.watch54.displays.Plugin
+version: 1.0.0
